[Release] sdk/resourcemanager/support/armsupport/2.0.0-beta.3 generation from spec commit: bcecb9990214471ee523e643da76a3a7c4170d2f

+++ sdk/resourcemanager/support/armsupport/filesnosubscription_client_example_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/support/armsupport/v2"
 )
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/9ec0fcc278aa2128c4fbb2b8a1aa93432d72cce0/specification/support/resource-manager/Microsoft.Support/preview/2022-09-01-preview/examples/ListFilesUnderFileWorkspace.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/bcecb9990214471ee523e643da76a3a7c4170d2f/specification/support/resource-manager/Microsoft.Support/preview/2023-06-01-preview/examples/ListFilesUnderFileWorkspace.json
 func ExampleFilesNoSubscriptionClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -47,10 +47,10 @@ func ExampleFilesNoSubscriptionClient_NewListPager() {
 		// 			Type: to.Ptr("Microsoft.Support/files"),
 		// 			ID: to.Ptr("/providers/Microsoft.Support/fileWorkspaces/testworkspace/files/test1.txt"),
 		// 			Properties: &armsupport.FileDetailsProperties{
-		// 				ChunkSize: to.Ptr[float32](41423),
+		// 				ChunkSize: to.Ptr[int32](41423),
 		// 				CreatedOn: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2022-08-24T20:18:19.000Z"); return t}()),
-		// 				FileSize: to.Ptr[float32](41423),
-		// 				NumberOfChunks: to.Ptr[float32](1),
+		// 				FileSize: to.Ptr[int32](41423),
+		// 				NumberOfChunks: to.Ptr[int32](1),
 		// 			},
 		// 		},
 		// 		{
@@ -58,17 +58,17 @@ func ExampleFilesNoSubscriptionClient_NewListPager() {
 		// 			Type: to.Ptr("Microsoft.Support/files"),
 		// 			ID: to.Ptr("/providers/Microsoft.Support/fileWorkspaces/testworkspace/files/test1.txt"),
 		// 			Properties: &armsupport.FileDetailsProperties{
-		// 				ChunkSize: to.Ptr[float32](41423),
+		// 				ChunkSize: to.Ptr[int32](41423),
 		// 				CreatedOn: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2022-08-24T20:18:19.000Z"); return t}()),
-		// 				FileSize: to.Ptr[float32](41423),
-		// 				NumberOfChunks: to.Ptr[float32](1),
+		// 				FileSize: to.Ptr[int32](41423),
+		// 				NumberOfChunks: to.Ptr[int32](1),
 		// 			},
 		// 	}},
 		// }
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/9ec0fcc278aa2128c4fbb2b8a1aa93432d72cce0/specification/support/resource-manager/Microsoft.Support/preview/2022-09-01-preview/examples/GetFileDetails.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/bcecb9990214471ee523e643da76a3a7c4170d2f/specification/support/resource-manager/Microsoft.Support/preview/2023-06-01-preview/examples/GetFileDetails.json
 func ExampleFilesNoSubscriptionClient_Get() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -91,15 +91,15 @@ func ExampleFilesNoSubscriptionClient_Get() {
 	// 	Type: to.Ptr("Microsoft.Support/files"),
 	// 	ID: to.Ptr("/providers/Microsoft.Support/fileWorkspaces/testworkspace/files/test.txt"),
 	// 	Properties: &armsupport.FileDetailsProperties{
-	// 		ChunkSize: to.Ptr[float32](41423),
+	// 		ChunkSize: to.Ptr[int32](41423),
 	// 		CreatedOn: to.Ptr(func() time.Time { t, _ := time.Parse(time.RFC3339Nano, "2022-08-24T20:18:19.000Z"); return t}()),
-	// 		FileSize: to.Ptr[float32](41423),
-	// 		NumberOfChunks: to.Ptr[float32](1),
+	// 		FileSize: to.Ptr[int32](41423),
+	// 		NumberOfChunks: to.Ptr[int32](1),
 	// 	},
 	// }
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/9ec0fcc278aa2128c4fbb2b8a1aa93432d72cce0/specification/support/resource-manager/Microsoft.Support/preview/2022-09-01-preview/examples/CreateFile.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/bcecb9990214471ee523e643da76a3a7c4170d2f/specification/support/resource-manager/Microsoft.Support/preview/2023-06-01-preview/examples/CreateFile.json
 func ExampleFilesNoSubscriptionClient_Create() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -112,9 +112,9 @@ func ExampleFilesNoSubscriptionClient_Create() {
 	}
 	_, err = clientFactory.NewFilesNoSubscriptionClient().Create(ctx, "testworkspace", "test.txt", armsupport.FileDetails{
 		Properties: &armsupport.FileDetailsProperties{
-			ChunkSize:      to.Ptr[float32](41423),
-			FileSize:       to.Ptr[float32](41423),
-			NumberOfChunks: to.Ptr[float32](1),
+			ChunkSize:      to.Ptr[int32](41423),
+			FileSize:       to.Ptr[int32](41423),
+			NumberOfChunks: to.Ptr[int32](1),
 		},
 	}, nil)
 	if err != nil {
@@ -122,7 +122,7 @@ func ExampleFilesNoSubscriptionClient_Create() {
 	}
 }
 
-// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/9ec0fcc278aa2128c4fbb2b8a1aa93432d72cce0/specification/support/resource-manager/Microsoft.Support/preview/2022-09-01-preview/examples/UploadFile.json
+// Generated from example definition: https://github.com/Azure/azure-rest-api-specs/blob/bcecb9990214471ee523e643da76a3a7c4170d2f/specification/support/resource-manager/Microsoft.Support/preview/2023-06-01-preview/examples/UploadFile.json
 func ExampleFilesNoSubscriptionClient_Upload() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
@@ -134,7 +134,7 @@ func ExampleFilesNoSubscriptionClient_Upload() {
 		log.Fatalf("failed to create client: %v", err)
 	}
 	_, err = clientFactory.NewFilesNoSubscriptionClient().Upload(ctx, "testworkspaceName", "test.txt", armsupport.UploadFile{
-		ChunkIndex: to.Ptr[float32](0),
+		ChunkIndex: to.Ptr[int32](0),
 		Content:    to.Ptr("iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAMAAAAoLQ9TAAAABGdBTUEAALGPC/xhBQAAAAFzUkdCAK7OHOkAAAAgY0hSTQAAeiYAAICEAAD6AAAAgOgAAHUwAADqYAAAOpgAABd"),
 	}, nil)
 	if err != nil {
